Removed previous path finder implementation
