+++ BTrainTests/Routing/PathFinderTests.swift
@@ -14,19 +14,8 @@ import XCTest
 @testable import BTrain
 
 class PathFinderTests: BTTestCase {
-        
+            
     func testSimplePath() throws {
-        let layout = LayoutECreator().newLayout()
-        let s1 = layout.block(for: Identifier<Block>(uuid: "s1"))!
-        let pf = PathFinder(layout: layout)
-        
-        let settings = PathFinder.Settings(random: false, reservedBlockBehavior: .avoidFirstReservedBlock, verbose: false)
-        let path = try pf.path(trainId: layout.trains[0].id, from: s1, direction: .next, settings: settings)
-        XCTAssertNotNil(path)
-        XCTAssertEqual(path!.description, ["s1:next", "b1:next", "b2:next", "b3:next", "s2:next"])
-    }
-    
-    func testSimplePath_New() throws {
         let layout = LayoutECreator().newLayout()
         let s1 = layout.block("s1")
         
@@ -35,26 +24,8 @@ class PathFinderTests: BTTestCase {
         XCTAssertEqual(path!.toBlockSteps.description, ["s1:next", "b1:next", "b2:next", "b3:next", "s2:next"])
         XCTAssertEqual(path!.toSteps.description, ["s1:next", "t1:(2>0)", "t2:(1>0)", "b1:next", "t3:(0>1)", "b2:next", "t4:(1>0)", "b3:next", "t5:(0>1)", "t6:(0>1)", "s2:next"])
     }
-
-    func testPathWithBacktrack() throws {
-        let layout = LayoutECreator().newLayout()
-        let s1 = layout.block(for: Identifier<Block>(uuid: "s1"))!
-        let pf = PathFinder(layout: layout)
-        pf.turnoutSocketSelectionOverride = { turnout, socketsId, context in
-            if turnout.id.uuid == "t5" {
-                return 2
-            } else {
-                return nil
-            }
-        }
-
-        let settings = PathFinder.Settings(random: false, reservedBlockBehavior: .avoidFirstReservedBlock, verbose: false)
-        let path = try pf.path(trainId: layout.trains[0].id, from: s1, direction: .next, settings: settings)
-        XCTAssertNotNil(path)
-        XCTAssertEqual(path!.description, ["s1:next", "b1:next", "b2:next", "b3:next", "b5:next", "b1:previous", "s2:previous"])
-    }
     
-    func testPathWithReservedTurnout_New() throws {
+    func testPathWithReservedTurnout() throws {
         let layout = LayoutECreator().newLayout()
         let s1 = layout.block("s1")
         let t6 = layout.turnout("t6")
@@ -66,29 +37,6 @@ class PathFinderTests: BTTestCase {
     }
     
     func testPathLookAhead() throws {
-        let layout = LayoutECreator().newLayout()
-        let s1 = layout.block(for: Identifier<Block>(uuid: "s1"))!
-        let b2 = layout.block(for: Identifier<Block>(uuid: "b2"))!
-        b2.reserved = Reservation("other", .next)
-        
-        let pf = PathFinder(layout: layout)
-
-        // Ensure that by specificy a look ahead equal to the number of blocks in the layout
-        // there is no valid path found because b2 is occupied.
-        let settings = PathFinder.Settings(random: false, reservedBlockBehavior: .avoidReserved, verbose: false)
-        var path = try pf.path(trainId: layout.trains[0].id, from: s1, direction: .next, settings: settings)
-        XCTAssertNil(path)
-        
-        // Now let's try again with a look ahead of just one block,
-        // in which case the reservation of b2 will be ignored because it is
-        // past the look ahead
-        let settings2 = PathFinder.Settings(random: false, reservedBlockBehavior: .avoidFirstReservedBlock, verbose: true)
-        path = try pf.path(trainId: layout.trains[0].id, from: s1, direction: .next, settings: settings2)
-        XCTAssertNotNil(path)
-        XCTAssertEqual(path!.description, ["s1:next", "b1:next", "b2:next", "b3:next", "s2:next"])
-    }
-
-    func testPathLookAhead_New() throws {
         let layout = LayoutECreator().newLayout()
         let s1 = layout.block("s1")
         let b2 = layout.block("b2")
@@ -107,28 +55,7 @@ class PathFinderTests: BTTestCase {
         XCTAssertEqual(path!.toBlockSteps.description, ["s1:next", "b1:next", "b2:next", "b3:next", "s2:next"])
     }
 
-    func testPathReserved() throws {
-        let layout = LayoutECreator().newLayout()
-        let s1 = layout.block(for: Identifier<Block>(uuid: "s1"))!
-        let b2 = layout.block(for: Identifier<Block>(uuid: "b2"))!
-        b2.reserved = Reservation("other", .next)
-        
-        let pf = PathFinder(layout: layout)
-        pf.turnoutSocketSelectionOverride = { turnout, socketsId, context in
-            if turnout.id.uuid == "t4" {
-                b2.reserved = nil
-            }
-            return nil
-        }
-
-        let settings = PathFinder.Settings(random: false, reservedBlockBehavior: .avoidReserved, verbose: false)
-        let path = try pf.path(trainId: layout.trains[0].id, from: s1, direction: .next, settings: settings)
-        XCTAssertNotNil(path)
-        XCTAssertFalse(path!.context.isOverflowing)
-        XCTAssertEqual(path!.description, ["s1:next", "b1:next", "b5:previous", "b3:previous", "b2:previous", "b1:previous", "s2:previous"])
-    }
-
-    func testPathWithReservedBlock_New() throws {
+    func testPathWithReservedBlock() throws {
         let layout = LayoutICreator().newLayout()
         let s1 = layout.block("s1")
         let b1 = layout.block("b1")
@@ -145,27 +72,6 @@ class PathFinderTests: BTTestCase {
     }
 
     func testPathBlockDisabled() throws {
-        let layout = LayoutECreator().newLayout()
-        let s1 = layout.block(for: Identifier<Block>(uuid: "s1"))!
-        let b2 = layout.block(for: Identifier<Block>(uuid: "b2"))!
-        b2.enabled = false
-        
-        let pf = PathFinder(layout: layout)
-        pf.turnoutSocketSelectionOverride = { turnout, socketsId, context in
-            if turnout.id.uuid == "t4" {
-                b2.enabled = true
-            }
-            return nil
-        }
-
-        let settings = PathFinder.Settings(random: false, reservedBlockBehavior: .avoidReserved, verbose: false)
-        let path = try pf.path(trainId: layout.trains[0].id, from: s1, direction: .next, settings: settings)
-        XCTAssertNotNil(path)
-        XCTAssertFalse(path!.context.isOverflowing)
-        XCTAssertEqual(path!.description, ["s1:next", "b1:next", "b5:previous", "b3:previous", "b2:previous", "b1:previous", "s2:previous"])
-    }
-
-    func testPathBlockDisabled_New() throws {
         let layout = LayoutICreator().newLayout()
         let s1 = layout.block("s1")
 
@@ -180,27 +86,6 @@ class PathFinderTests: BTTestCase {
     }
 
     func testPathTurnoutDisabled() throws {
-        let layout = LayoutECreator().newLayout()
-        let s1 = layout.block(for: Identifier<Block>(uuid: "s1"))!
-        let t4 = layout.turnout(for: Identifier<Turnout>(uuid: "t4"))!
-        t4.enabled = false
-        
-        let pf = PathFinder(layout: layout)
-        pf.turnoutSocketSelectionOverride = { turnout, socketsId, context in
-            if turnout.id.uuid == "t5" {
-                t4.enabled = true
-            }
-            return nil
-        }
-
-        let settings = PathFinder.Settings(random: false, reservedBlockBehavior: .avoidReserved, verbose: false)
-        let path = try pf.path(trainId: layout.trains[0].id, from: s1, direction: .next, settings: settings)
-        XCTAssertNotNil(path)
-        XCTAssertFalse(path!.context.isOverflowing)
-        XCTAssertEqual(path!.description, ["s1:next", "b1:next", "b5:previous", "b3:previous", "b2:previous", "b1:previous", "s2:previous"])
-    }
-
-    func testPathTurnoutDisabled_New() throws {
         let layout = LayoutICreator().newLayout()
         let s1 = layout.block("s1")
 
@@ -220,35 +105,16 @@ class PathFinderTests: BTTestCase {
         layout.reserve("NE1", with: "1", direction: .next)
         
         let train = layout.trains[0]
-        let currentBlock = layout.block(for: Identifier<Block>(uuid: "NE1"))!
-        let destination = Destination(Identifier<Block>(uuid: "LCF1"))
-        
-        let pf = PathFinder(layout: layout)
-        let settings = PathFinder.Settings(random: false, reservedBlockBehavior: .avoidFirstReservedBlock, verbose: false)
-        let path = try pf.path(trainId: train.id, from: currentBlock, destination: destination, direction: .next, settings: settings)
-        XCTAssertNotNil(path)
-        XCTAssertEqual(path!.description, ["NE1:next", "OL1:next", "OL2:next", "OL3:next", "NE4:next", "IL1:next", "IL2:next", "IL3:next", "S:next", "IL1:previous", "IL4:previous", "IL3:previous", "IL2:previous", "OL1:previous", "NE3:previous", "M1:next", "M2U:next", "LCF1:next"])
-        
-        self.measure {
-            let path = try! pf.path(trainId: train.id, from: currentBlock, destination: destination, direction: .next, settings: settings)
-            XCTAssertNotNil(path)
-        }
-    }
-
-    func testPathBetweenStations_New() throws {
-        let layout = LayoutFCreator().newLayout()
-        
-        layout.reserve("NE1", with: "1", direction: .next)
-        
-        let train = layout.trains[0]
         let ne1 = layout.block("NE1")
         let lcf1 = layout.block("LCF1")
 
-        let path = layout.path(for: train, from: (ne1, .next), to: (lcf1, .next), reservedBlockBehavior: .avoidFirstReservedBlock)
-        XCTAssertEqual(path!.toBlockSteps.description, ["NE1:next", "OL1:next", "OL2:next", "OL3:next", "NE4:next", "IL1:next", "IL2:next", "IL3:next", "S:next", "IL1:previous", "IL4:previous", "IL3:previous", "IL2:previous", "OL1:previous", "NE3:previous", "M1:next", "M2U:next", "LCF1:next"])
+        let path = layout.path(for: train, from: (ne1, .next), to: (lcf1, .next), reservedBlockBehavior: .avoidFirstReservedBlock)!
+        XCTAssertEqual(path.toBlockSteps.description, ["NE1:next", "OL1:next", "OL2:next", "OL3:next", "NE4:next", "IL1:next", "IL2:next", "IL3:next", "S:next", "IL1:previous", "IL4:previous", "IL3:previous", "IL2:previous", "OL1:previous", "NE3:previous", "M1:next", "M2U:next", "LCF1:next"])
         
-        // TODO: try with settings random=true and this time the shortest path should be returned
-        
+        // Note: by enabling random routing and specifying a destination, the shortest path will be returned
+        let shortestPath = layout.path(for: train, from: (ne1, .next), to: (lcf1, .next), reservedBlockBehavior: .avoidFirstReservedBlock, random: true)!
+        XCTAssertTrue(shortestPath.toBlockSteps.count < path.toBlockSteps.count)
+
         self.measure {
             let path = layout.path(for: train, from: (ne1, .next), to: (lcf1, .next), reservedBlockBehavior: .avoidFirstReservedBlock)
             XCTAssertNotNil(path)
@@ -261,57 +127,15 @@ class PathFinderTests: BTTestCase {
         layout.reserve("s1", with: "1", direction: .next)
         
         let train = layout.trains[0]
-        let currentBlock = layout.block(for: Identifier<Block>(uuid: "s1"))!
-        let destination = Destination(Identifier<Block>(uuid: "s2"))
-
-        let pf = PathFinder(layout: layout)
-        
-        let settings = PathFinder.Settings(random: false, reservedBlockBehavior: .avoidFirstReservedBlock, verbose: false)
-        let path = try pf.path(trainId: train.id, from: currentBlock, destination: destination, direction: .next, settings: settings)
-        XCTAssertEqual(path!.description, ["s1:next", "b1:next", "b2:next", "b3:next", "s2:next"])
-    }
-
-    func testPathBetweenStations2_New() throws {
-        let layout = LayoutECreator().newLayout()
-        
-        layout.reserve("s1", with: "1", direction: .next)
-        
-        let train = layout.trains[0]
         let s1 = layout.block("s1")
         let s2 = layout.block("s2")
 
         let path = layout.path(for: train, from: (s1, .next), to: (s2, .next), reservedBlockBehavior: .avoidFirstReservedBlock)
         XCTAssertEqual(path!.toBlockSteps.description, ["s1:next", "b1:next", "b2:next", "b3:next", "s2:next"])
     }
-
-    func testShortestPathBetweenStations() throws {
-        let layout = LayoutFCreator().newLayout()
-        
-        layout.reserve("NE1", with: "1", direction: .next)
-        
-        let train = layout.trains[0]
-        let currentBlock = layout.block(for: Identifier<Block>(uuid: "NE1"))!
-        let destination = Destination(Identifier<Block>(uuid: "HLS_P1"))
-
-        let pf = PathFinder(layout: layout)
-        let settings = PathFinder.Settings(random: true, reservedBlockBehavior: .avoidFirstReservedBlock, verbose: false)
-        
-        var generatedPaths = [PathFinder.Path]()
-        let path = try pf.path(trainId: train.id, from: currentBlock, destination: destination, direction: .next, settings: settings) { path in
-            generatedPaths.append(path)
-        }
-        XCTAssertNotNil(path)
-        XCTAssertEqual(generatedPaths.count, 10)
-        
-        generatedPaths.sort { p1, p2 in
-            p1.steps.count < p2.steps.count
-        }
-        let shortestPath = generatedPaths[0]
-        XCTAssertEqual(path!.description, shortestPath.description)
-    }
     
     // TODO: investigate why this is so slow
-    func testShortestPathBetweenStations_New() throws {
+    func testShortestPathBetweenStations() throws {
         let layout = LayoutFCreator().newLayout()
         
         layout.reserve("NE1", with: "1", direction: .next)
@@ -340,13 +164,13 @@ extension Layout {
         self.block(for: Identifier<Block>(uuid: block))?.reserved = nil
     }
     
-    func path(for train: Train, from: (Block, Direction), to: (Block, Direction)?, reservedBlockBehavior: PathFinder.Settings.ReservedBlockBehavior = .avoidReserved) -> GraphPath? {
-        let settings = LayoutPathFinder.Settings(reservedBlockBehavior: reservedBlockBehavior, baseSettings: GraphPathFinder.Settings(verbose: false, random: false, overflow: pathFinderOverflowLimit))
+    func path(for train: Train, from: (Block, Direction), to: (Block, Direction)?, reservedBlockBehavior: LayoutPathFinder.ReservedBlockBehavior = .avoidReserved, random: Bool = false) -> GraphPath? {
+        let settings = LayoutPathFinder.Settings(reservedBlockBehavior: reservedBlockBehavior, baseSettings: GraphPathFinder.Settings(verbose: false, random: random, overflow: pathFinderOverflowLimit))
         let gl = LayoutPathFinder(layout: self, train: train, settings: settings)
         return path(for: train, from: from, to: to, pathFinder: gl, constraints: gl.constraints)
     }
     
-    func shortestPaths(for train: Train, from: (Block, Direction), to: (Block, Direction)?, reservedBlockBehavior: PathFinder.Settings.ReservedBlockBehavior = .avoidReserved) -> [GraphPath] {
+    func shortestPaths(for train: Train, from: (Block, Direction), to: (Block, Direction)?, reservedBlockBehavior: LayoutPathFinder.ReservedBlockBehavior = .avoidReserved) -> [GraphPath] {
         let settings = LayoutPathFinder.Settings(reservedBlockBehavior: reservedBlockBehavior, baseSettings: GraphPathFinder.Settings(verbose: false, random: true, overflow: pathFinderOverflowLimit))
         let gl = LayoutPathFinder(layout: self, train: train, settings: settings)
         
@@ -369,12 +193,4 @@ extension Array where Element == Route.Step {
     var description: [String] {
         return self.map { $0.description }
     }
-}
-
-extension PathFinder.Path {
-        
-    var description: [String] {
-        return steps.description
-    }
-    
 }
